Changing application configuration to .conf

+++ src/main/kotlin/org/kryptokrona/api/services/HashrateService.kt
@@ -36,7 +36,7 @@ interface HashrateService {
 
     fun getAll()
 
-    fun getById(id: Long): Hashrate?
+    fun getById(id: Int): Hashrate?
 
     fun save(hashrate: Hashrate)
